feat/auction: 내 판매 경매 조회 및 상세 조회에 AuthUser, BidService 연동 (connect AuthUser & BidService for my selling auction list/detail APIs)

+++ src/main/java/org/example/lastcall/domain/auction/dto/response/MySellingResponse.java
@@ -15,18 +15,13 @@ public class MySellingResponse {
     private String imageUrl;
     private String productName;
     private String productDescription;
-    private int currentBid;
+    private Long currentBid;
     private AuctionStatus status;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
 
     // 정적 팩토리 메서드 (from)
-    public static MySellingResponse from(
-            Auction auction,
-            Product product,
-            String imageUrl,
-            int currentBid
-    ) {
+    public static MySellingResponse from(Auction auction, Product product, String imageUrl, Long currentBid) {
         return MySellingResponse.builder()
                 .id(auction.getId())
                 .imageUrl(imageUrl)
